requirements.txt
cachetools==4.2.1
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
dnspython==2.1.0
ecdsa==0.16.1
email-validator==1.1.2
Flask==1.1.2
Flask-Cors==3.0.10
flask-mongoengine==1.0.0
Flask-Script==2.0.6
Flask-WTF==0.14.3
google-auth==1.28.0
gunicorn==20.0.4
hashids==1.2.0
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1
mock==3.0.5
mongoengine==0.23.0
Pillow==6.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pymongo==3.11.3
python-dotenv==0.15.0
python-jose==3.1.0
requests==2.25.1
rsa==4.7.2
six==1.15.0
urllib3==1.26.4
Werkzeug==1.0.1
WTForms==2.3.3
